Move source files to /src/blend2d, support working text, add text example, add C binding information to README

+++ src/blend2d/text/font.cr
@@ -4,7 +4,7 @@ module Blend2D::Text
     class Font < BLStructure
         @core = uninitialized LibBlend2D::BLFontCore
 
-        def initialize(face : FontFace, size : Float64)
+        def initialize(face : FontFace, size : Float32)
             LibBlend2D.font_init(pointer)
             LibBlend2D.font_create_from_face(pointer, face.pointer, size)
         end
